Add implementation, README, test runner and unit tests
The project has been developed in an unrelated sandbox repository. This
commit adds the bulk of that work. The implementation lacks only one
feature that I think should be in the first release: ability to edit
the commit message interactively. That is, when reporting error(s)
in commit-msg hook the script should ask
  Continue anyway? [yes/no/edit]
instead of only
  Continue anyway? [yes/no]

README contains essentially everything that is needed in the first
version. Still need to add a logo and some badges.

The test framework (testrunner and testlib) is not too polished but
effective. I probably need to copy simpleargs library to this repository
so that the tests can be run without first obtaining simpleargs.

The unit tests are comprehensive but unfortunately there's no code
coverage analysis... yet?