Add explicit author to git commit
Add --author flag to 'git commit' that's used inside lint-commit-msg
for commit message cleanup. This is done because the tests are
failing in Github workflow with the "Author identity unknown" error.
I'm not sure if this would ever be a problem in real usage but let's
see if the tests start passing in Github workflow.