Fix bug: 'suppress' is a verb in imperative mood
The old logic reported "subject not in imperative mood" error if
the subject's first word ended in 's' (e.g. "updates"). However,
that simple heuristic did not take into account verbs like 'suppress'
that end in 's' in their imperative mood. So, only flag verbs that
end in 's' but not 'ss'.