Make lint-commit-msg work with Egit (Eclipse Git plugin)
This changeset fixes a bug where making a commit from Eclipse IDE
with lint-commit-msg included in the commit-msg hook resulted in
lint-commit-msg being called recursively. This resulted in the commit
operation getting stuck for dozens of seconds and finally giving an
error message.

The problem was that when a commit is made in Eclipse lint-commit-msg
is called with environment variables GIT_DIR and GIT_WORK_TREE set
to .git directory and the repository directory respectively. From line
275 onward lint-commit-msg uses 'git' commands to cleanup the commit
message before linting it (see the code comments for detailed
explanation why this is done). However, with GIT_DIR and GIT_WORK_TREE
set the git operations were targeting the user's repository and not
the temporary repository created for performing the commit message
cleanup.

Now, because as part of the cleanup 'git commit' is called (line 284)
lint-commit-msg is executed again (as part of the commit-msg hook) and
again and again (recursively)... So, fix this by unsetting the
environment variables prefixed with "GIT" so that the git commands used
within lint-commit-msg are operating on the temporary repository.

Also, remove some redundant semicolons and complement tests.