Use interactive mode only if stdout is terminal
Earlier if LCM_INTERACTIVE was set to 'auto' interactive mode was
turned on if lint-commit-msg was being run within commit-msg hook.
This logic was flawed. If the user makes a commit from within an IDE
(e.g. Eclipse) we don't want to use interactive mode by default because
it freezes the IDE. (The IDE keeps waiting for user input if there
are linting errors).

Therefore, separate the two conditions: "is committing" and "stdout is
terminal". Now automatically turning on interactive mode (i.e. asking
the user "Continue anyways?") requires that we are committing (i.e. we
are inside commit-msg hook) AND stdout is terminal. Thus, the two
scenarios where we shouldn't use interactive mode are handled correctly:
1. When we're running lint-commit-msg "standalone" (not within
   commit-msg hook). There's no use asking the user "Continue anyway?"
   because there's nothing to continue even if the user replies "yes".
2. When lint-commit-msg is run in an environment where stdout (and
   probably stderr and stdin) is not a terminal e.g. when committing
   from within an IDE.