Print usage instructions before checking for invalid env vars
If the user runs 'lint-commit-msg -h' we should always print the usage
instructions... even in the case that the user has e.g. a misspelled
configuration variable (e.g. LCM_INTREACTIVE=never) in their .bashrc.

The same rationale applies with LCM_IGNORE_ALL=true, it should bypass
the linting even if there are errors in the configuration variables.