Print (error) messages to stderr instead of stdout
First of all, it seems that everything printed in commit-msg hook
end up in stderr anyways. In that sense this change affects nothing.
When run standalone (not in commit-msg hook) the change is not apparent
either because stdout and stderr are usually the same thing i.e. the
user's terminal.

So, why this change? It seems that Eclipse IDE (or its Egit plugin)
calls the commit-msg hook differently to how it works when running
'git commit' on command line: When the user tries to commit in Eclipse
and there are linting errors a dialog pops up with a title saying
  Commit was aborted by hook "commit-msg"
and the dialog body contains what commit-msg hook has printed to stderr.
Output written to stdout is not visible in the UI. (It seems to be
visible in the terminal if Eclipse was launched from the terminal as
a foreground process.)

This change affects almost all of the test cases that now need to
inspect the contents of stderr instead of stdout. This causes also
other complexities: In the tests 'unbuffer' is used to trick
lint-commit-msg to believe that its stdout is a terminal (in order
to test e.g. colored output). 'expect' command is used to test the
interactive parts i.e. simulating the user replying "yes" or "no".
(Unfortunately) both tools also make all the output from lint-commit-msg
go to stdout. So, tests using either 'unbuffer' or 'expect' need
to inspect stdout although in reality lint-commit-msg now always prints
its messages (apart from usage instructions) to stderr.