Fix false "subject not capitalized" errors (fixes #1)
Fix bug that on some environments caused valid commit messages being
reported as having a non-capitalized subject line. For example, subject
line such as "Version bump 1.0.0" was reported as not being capitalized
because the regular expression used a character range [a-z] that on
some systems is interpreted as containing uppercase characters as well
(depends on LC_COLLATE environment variable).