Add _LCM_TRACE env variable for debugging
Setting _LCM_TRACE=true causes 'set -x' to be run as the first thing
that lint-commit-msg does. The script already has _LCM_DEBUG that can
be set to 'true' to get some diagnostics output. However, debugging
a bug (occurring possibly only on someone else's system) will be much
easier when one can ask the reporter (of the bug) to run the command
again with _LCM_TRACE=true and to include the output in the bug report.

Also, print a helpful error message if someone sets LCM_DEBUG=true
or LCM_TRACE=true (the correct variables start with an underscore).