Make capitalization check work with 'nocasematch' on
Bash has an option, 'nocasematch' (see 'help shopt') that makes pattern
matching case-insensitive. So, the check for "subject line capitalized"
  [[ ${subject} =~ ^[[:lower:]] ]]
will report a non-capitalized subject whenever the subject line starts
with an ASCII letter, lowercase or uppercase, that is, basically always.

As far as I know, 'nocasematch' is rarely on. However, if it is on,
without this fix it would produce a very confusing error message.
For example,

$ git commit -m "Add HTTP client"
ERROR: commit message not properly formatted
- line 1: subject not capitalized
  Add HTTP client