Fix mood check when verb is all uppercase
Consider commit message subject line
  UPDATED configuration parsing
Before this fix the error message would have been
  - line 1: verb 'UPDATED' in the subject not in imperative mood,
            use for example 'Add' instead of 'AddED'
Notice the last 'ED' which is now fixed to 'ed'. Of course the words
in the commit message shouldn't be in all uppercase but nevertheless
this commit fixes one more loophole.

Also, because of the possibly unexpected consequences of LC_COLLATE,
changed some character ranges to safer ones e.g. [a-z] -> [:lower:]